BlockHeader refactored

+++ bsv/src/bitcoin/encoding.rs
@@ -11,37 +11,18 @@ use tokio::io::{AsyncRead, AsyncWrite};
 ///
 /// This is the non-async trait for reading and writing to buffers. See [AsyncEncodable] for the
 /// async trait to read and write with truly async sources.
-///
-/// The advantage of using async sources is that you can read and write data a byte at a time. If the
-/// data you are reading is enormous, this can be very beneficial because you can handle the data as
-/// it arrives, possibly without storing it. This is particularly useful in a hostile environment that
-/// allows huge data.
-///
-/// However, lets be frank, its a pain in the neck. Doing this properly requires handling the incoming
-/// data in a streaming fashion all the way up the stack. Handling objects (transactions, etc) entirely
-/// in memory is much easier and arguably performs better, particularly when you can allocate a
-/// contiguous space in memory to hold the entire object, instead of having to allocate multiple
-/// sections for the various parts. Also note that the Bitcoin P2P protocol is message based,
-/// with the message size included in the header, so an entire message can be fetched at once before
-/// processing.
-///
-/// So, there are two traits for encoding and decoding Bitcoin structures to and from binary. This is
-/// the non-async trait which makes use of [bytes::Bytes] to avoid copying memory around. The async
-/// trait is [AsyncEncodable]. If you don't need the async capabilities, use this one.
 pub trait Encodable {
     /// Read the data structure from a buffer.
-    fn from_binary(buffer: &mut dyn Buf) -> Result<Self>
-    where
-        Self: Sized;
+    fn from_binary(buffer: &mut dyn Buf) -> Result<Self> where Self: Sized;
 
     /// Write the data structure to a buffer.
     fn to_binary(&self, buffer: &mut dyn BufMut) -> Result<()>;
 
     /// Return the size of the serialized form.
-    // It is vital that implementations of this function use a method that does not just serialize the object
+    // It is vital (for efficiency) that implementations of this function use a method that does not just serialize the object
     // and count the bytes. This is because this function is used to determine the size of the buffer to allocate
     // for the serialization.
-    fn size(&self) -> usize;
+    fn size(&self) -> u64;
 }
 
 /// Asynchronously read & write Bitcoin data structures to and from binary in Bitcoin encoding format.
@@ -65,7 +46,7 @@ pub trait AsyncEncodable {
     // It is vital that implementations of this function use a method that does not just serialize the object
     // and count the bytes. This is because this function is used to determine the size of the buffer to allocate
     // for the serialization.
-    fn async_size(&self) -> usize;
+    fn async_size(&self) -> u64;
 
     /// Read the data structure from a buffer.
     ///
